requirements.txt
Django==4.2.7
python-docx==0.8.11
polars==0.20.2
openpyxl==3.1.2
# AI integration - Flan-T5
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.25.0
python-decouple==3.8
requests==2.31.0
